Add fix according to required changes
