+++ index.html
@@ -54,6 +54,6 @@
     </main>
     <footer class="copyright">Copyright.io
     </footer>
-    <script type="module" src="./modules/index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
