--- modules/index.js
@@ -1,117 +1,0 @@
-import {
-  titlei, authori, bookForm, formSection, toAdd, contactInfo, contactSection,
-  mainHead, booksSection, checkList, dateTime,
-} from './htmlElements.js';
-import luxon from './luxon.js';
-
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-
-  populateLocalStorage = () => {
-    const books = this.getBooks();
-    if (!books) {
-      localStorage.setItem('books', JSON.stringify([]));
-    }
-  };
-
-  getBooks = () => {
-    const books = JSON.parse(localStorage.getItem('books'));
-    return books || [];
-  };
-
-  addBook = (book) => {
-    const books = this.getBooks();
-    books.push(book);
-    localStorage.setItem('books', JSON.stringify(books));
-  };
-
-  removeBookFromLocalStorage = (id) => {
-    this.removeBook(id);
-    this.showBook();
-  };
-
-  removeBook = (id) => {
-    const books = this.getBooks();
-    const newBooks = books.filter((book) => book.id.toString() !== id);
-    localStorage.setItem('books', JSON.stringify(newBooks));
-    document.querySelector(`#container${id}`).remove();
-  };
-
-  showBook = () => {
-    this.populateLocalStorage();
-    const books = this.getBooks();
-    const booksSection = document.getElementById('books');
-    const booksList = document.createElement('div');
-    booksList.className = 'list';
-    let booksElement = '';
-    for (let i = 0; i < books.length; i += 1) {
-      booksElement = `
-      <div id="container${books[i].id}" class="booksList">
-              <div class="inputs">
-              <p id="display-title">${books[i].title}</p>
-              <p id="display-author">${books[i].author}</p>
-              </div>
-              <button type="button" id="${books[i].id}" class ="remove"
-              >Remove</button>
-            </div>
-            <hr>
-    `;
-      booksSection.appendChild(booksList);
-      booksList.innerHTML += booksElement;
-      const remove = document.querySelectorAll('.remove');
-      if (remove.length) {
-        remove.forEach((button) => {
-          button.addEventListener('click', () => {
-            this.removeBookFromLocalStorage(button.id);
-            window.location.reload();
-          });
-        });
-      }
-    }
-  };
-}
-const book = new Book(titlei.value, authori.value);
-bookForm.addEventListener('submit', (e) => {
-  e.preventDefault();
-  const book = new Book(titlei.value, authori.value);
-  const { title, author } = book;
-  const id = Math.round(Math.random() * 10000000);
-  const newObject = {
-    id,
-    title,
-    author,
-  };
-  if (title && author) {
-    book.addBook(newObject);
-    bookForm.reset();
-    book.showBook();
-  }
-  window.location.reload();
-});
-window.addEventListener('DOMContentLoaded', book.showBook());
-
-toAdd.addEventListener('click', () => {
-  formSection.style.display = 'block';
-  booksSection.style.display = 'none';
-  checkList.style.color = 'black';
-  toAdd.style.color = 'blue';
-  contactInfo.style.color = 'black';
-  contactSection.style.display = 'none';
-  mainHead.style.display = 'none';
-});
-checkList.addEventListener('click', () => {
-  window.location.reload();
-});
-contactInfo.addEventListener('click', () => {
-  booksSection.style.display = 'none';
-  contactSection.style.display = 'block';
-  contactInfo.style.color = 'blue';
-  checkList.style.color = 'black';
-  formSection.style.display = 'none';
-  toAdd.style.color = 'black';
-  mainHead.style.display = 'none';
-});
-dateTime.innerText = luxon.DateTime.now().setLocale('en-US').toLocaleString(luxon.DateTime.DATETIME_FULL_WITH_SECONDS);
